requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
et-xmlfile==1.1.0
Flask==2.1.3
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
openai==0.22.0
openpyxl==3.0.10
pandas==1.4.3
pandas-stubs==1.4.3.220724
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
tqdm==4.64.0
types-pytz==2022.1.2
urllib3==1.26.11
Werkzeug==2.2.1
